make callable stlascii2stlbin script
