+++ api/pylele_utils.py
@@ -225,30 +225,6 @@ def make_or_exist_path(out_path):
 
     assert os.path.isdir(out_path), f"Cannot export to non directory: {out_path}"
 
-def stl_is_bin(fname) -> bool:
-    """ Returns True if .scl in ascii format """
-    with open(fname, "r",encoding='utf-8') as fp:
-        retval=fp.readline().find("solid") > -1
-        fp.close()
-    return retval
-
-def stl2bin(infile,outfile='') -> str:
-    """ Converts an ASCII .stl into a binary """
-    assert os.path.isfile(infile), f"ERROR: Input File {infile} does not exist!"
-
-    if not stl_is_bin(infile):
-        if outfile=='':
-            fname,fext = os.path.splitext(infile)
-            outfile = f'{fname}_bin{fext}'
-
-        cmdstr = f'stl2bin {infile} {outfile}'
-        os.system(cmdstr)
-        assert os.path.isfile(outfile), f"ERROR: Output File {outfile} does not exist!"
-        return outfile
-    # else
-    print(f'WARNING: .stl {infile} is already in binary format!')
-    return infile
-
 def gen_stl_foo(outpath: str, bin_en = True) -> None:
     """ generate an .stl file """
 
@@ -358,7 +334,7 @@ def gen_stl_foo(outpath: str, bin_en = True) -> None:
     return fout
 
 def gen_scad_foo(outpath: str, module_en = True) -> None:
-    """ generate an .stl file """
+    """ generate a .scad file """
 
     if module_en:
         stlstr="""
